# What is the problem
- When we make doxygen output , each plantuml comment spends a lot of time (more than 1 second).
	- If doxygen generates 6 plantuml , it takes 18.124 sec (98%). But total time is 18.498 sec (100%).
	- Generating plantuml on java occupied most of time.
- It regenerate all plantuml output even though we change or add just one plantuml comments  or we change the doxygen comments.
	- Reuse the pre-generated plantuml output!!

# Main Idea
- Goal
	- Reduce the running time of doxygen from making plantuml result.
- Method
	- Reuse the result(png or svg ...) of unchanged plantuml comments.

```
9.689 sec: Total elapsed time: 9.689 seconds
(of which 9.340 seconds waiting for external tools to finish)
```
- According to upper timeline ,  generating images with java is a big burden.
- So I try to remove java running process.   I guess we can reduce the java running time if we reuse already generated images.

# What to do
- add configuration value in Doxyfile
	- PLANTUML_CACHE = YES
	- PLANTUML_CACHE_DEBUG_PRINT = YES			<- It has dependency on PLANTUML_CACHE
- We reduced the time to run the java process for making the plantuml result (png, svg ...).
- Changed Configuration (Doxyfile)
	- HAVE_DOT               = YES
	- DOT_IMAGE_FORMAT       = png
	- GENERATE_HTML          = YES
		- generate  png  file from plantuml
	- GENERATE_DOCBOOK       = YES
		- generate  png  file from plantuml
	- GENERATE_PERLMOD       = YES
		- make a perl module
- results
	- Original : original doxygen makes html and docbook plantuml png separately. (my example has 3 plantuml comments in source code.)
	- First Trial [^1] : When we use modified version , html is running first and then reuse the html's plantuml png results for generating docbook.
		- reduce the time to make a docbook plantuml result.
	- Second Trial [^2] : Change One comment : When we change one of three plantuml comments, we can reuse the previous compiled doxygen result.
		- reduce the time to make a plantuml for unchanged plantuml.
	- Change output type [^3] : When we change output format png into svg, we can resue the previous compiled doxygen result about same contents..
		- DOT_IMAGE_FORMAT       = png  -> svg
		- reduce the time to make a plantuml for unchanged plantuml with the same format..

|        Test Type          |  Original  |  First [^1] | Change One [^2] | Change output type [^3] |
|---------------------------|------------|-------------|-----------------|-------------------------|
| Elapsed Time              | 18.498 sec | 9.689 sec   |  3.153 sec      |  9.744 sec              |
| Guessing Time (Only HTML) | 9.498 sec  | 9.498 sec   |  3.153 sec      |  9.744 sec              |

- Analysis from results
	- Although you use only html type ,  you can get the benefit from second trials.


[^1] : First - When we use modified version , html is running first and then reuse the html's plantuml png results for generating docbook.
[^2] : Change One - When we change one of three plantuml comments, we can reuse the previous compiled doxygen result.
[^3] : Change output type - When we change output format png into svg, we can reuse the previous compiled doxygen result about same contents..