@albert-github 2018-08-24
The caching of the information in one way or the other is a good idea.
When generating the the `dot ` images the `.dot` files are written and based on this `.dot` files a `MD5` sum is calculated. With subsequent calculations the `.dot` file is written as well as the `MD5` sum is calculated. when the `MD5` sums are identical the image is not regenerated. Furthermore with `dot` images these are generated in multiple threads.
Another possibility would be to collect all plantuml requests in one file and process this file. Plantuml will generated a new image for each original plantuml request. Something similar is done for formulas. This gives less overhead as only 1 process is started.

I think one of these techniques would be sufficient.


Regarding the code (small review):

- in case of debug you use printf statements, I think that here there better options would be
  - create an -d flag (e.g. -d plantuml) that gives the information, so no need for using the configuration option
 - usage of printf
   - printf should not be used,all output is steered over calls to routines as defined in message.cpp, 
 - usage of std::string / std::map
    - in doxygen the string manipulation is. mostly, done by means of Qt classes (e.g. classes QCString, QString). For consistency these classes should be used.
    - map manipulation is also done by means of Qt classes )dictionary and list classes)
 - dirent / opendir /readdir
    - when  browsing through directories the routines line `readFileOrDirectory` should probably be used or the techniques used in this routine