@trec789 2019-10-27
Hello cheoljoo,

I believe I am having some issues with this change.
When using Doxygen 1.8.16 with html output the plantuml images won't update. I know that it is the intention but it causes the following problems:
- if I remove the html files in the output folder (for whatever reason) the plantuml images are not recreated.
- if I change the PLANTUML_CFG_FILE config param in Doxyfile, e.g. for changing the diagrams color, the plantuml images won't be updated.

The issues seems to be related to Doxygen 1.8.16 only. Doxygen 1.8.15 works fine since it always re-create the images.
I tried to add PLANTUML_CACHE = NO without success.

My system:
Windows 10
Doxygen 1.8.16
plantuml.jar 1.2019.11
Graphviz 2.38.0