@albert-github 2019-10-27
Do you mean that you removed the individual html files or the complete html directory ?

A workaround will be to remove also the file `inline_umlgraph_cache_all.pu`, probably it is in the directory where you started doxygen.