@albert-github 2019-10-27
Problems like noted with the `PLANTUML_CFG_FILE` will occur when:
- `PLANTUML_JAR_PATH`, the jar file is replaced for another / new one (name does not have to change though!)
- `PLANTUML_INCLUDE_PATH` , when the content changes or the contents of the file. This is quite difficult to detect as an include can include again files from other places.

**Note** this is not the pull request that landed into release but the pull request #6723 did. The setting `PLANTUML_CACHE` does not exist here anymore