@albert-github 2019-10-27
@trecmar please open a new issue with the mentioned problems as otherwise the problems might get lost. **Don't reopen** this pull request (or the #6723 pul request) for this purpose.