@trecmar 2019-10-27
Hi,

Regarding the first email:
- I removed the individual html files only, not the html dir.
- Removing the inline_umlgraph_cache_all.pu worked. Thank you for the tip.

Second email:
- I agree that change the configuration causes a change in the output, but is difficult to detect.
- I would suggest adding some kind of Doxygen command line flag to force a “clean build” or “rebuild all” type of thing.

Third email:
- I will open a new issue as suggested and add this info.

Thanks,
Marc



From: albert-github
Sent: Sonntag, 27. Oktober 2019 10:27
To: doxygen/doxygen
Cc: trecmar; Mention
Subject: Re: [doxygen/doxygen] Dramatically reduce the time when rerun thedoxygen with plantuml comments (Imporvement of performance) (#6465)

@trecmar please open a new issue with the mentioned problems as otherwise the problems might get lost. Don't reopen this pull request (or the #6723 pul request) for this purpose.
—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub, or unsubscribe.