@cheoljoo 2019-10-29
sorry @trecmar  for my late response.  albert gave the answer.  thanks @albert-github 
when you remove html files in html direcotry , you should remove inline_umlgraph_cache_all.pu file also.
because doxygen will decide after comparison with inline_umlgraph_cache_all.pu whether doxygen create the plantuml png files or not.

and I will try to add some configure options in Doxyfile. but doxygen board members decide that this is default.   because we can find the bug easily when users use this option as default.

Thanks all.  cheoljoo.